+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 * Fixed FoldedTrinket+ text
 * Fixed BandageBot: now stacks correctly
 * Fixed typos: FlamableWall -> FlammableWall; FlamableTrinket -> FlammableTrinket
+* Fixed: BandageBot and Assemble will create upgraded trinkets if you have Polishing Wheel
 * Changed Generator: now costs 1(0), instead of 2(1)
 * Changed Now or Never: gives 11(14) block next turn instead of 12(16)
 * Removed Nitro Boost, it was never a collectible card to begin with